backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
pytz==2025.2
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
